requirements.txt
aspose-zip==24.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
filelock==3.9.0
fonttools==4.49.0
fsspec==2023.4.0
gdown==5.1.0
idna==3.6
imbalanced-learn==0.12.0
importlib_resources==6.3.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.24.1
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
soupsieve==2.5
sympy==1.12
threadpoolctl==3.3.0
torch==2.2.1+cu121
torchaudio==2.2.1+cu121
torchvision==0.17.1+cu121
tqdm==4.66.2
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.2.1
xgboost==2.0.3
zipp==3.18.0
